+++ crates/namada/src/ledger/native_vp/masp.rs
@@ -21,9 +21,9 @@ use namada_core::masp::{encode_asset_type, MaspEpoch};
 use namada_core::storage::Key;
 use namada_gas::GasMetering;
 use namada_governance::storage::is_proposal_accepted;
-use namada_sdk::masp::verify_shielded_tx;
 use namada_state::{ConversionState, OptionExt, ResultExt, StateRead};
 use namada_token::read_denom;
+use namada_token::validation::verify_shielded_tx;
 use namada_tx::BatchedTxRef;
 use namada_vp_env::VpEnv;
 use ripemd::Digest as RipemdDigest;
