+++ crates/shielded_token/Cargo.toml
@@ -15,17 +15,27 @@ version.workspace = true
 [features]
 default = []
 multicore = ["dep:rayon"]
-testing = ["multicore", "namada_core/testing"]
+# Download MASP params if they're not present
+download-params = ["masp_proofs/download-params"]
+testing = [
+  "multicore",
+  "namada_core/testing",
+  "masp_primitives/test-dependencies",
+]
 
 [dependencies]
 namada_controller = { path = "../controller" }
 namada_core = { path = "../core" }
+namada_gas = { path = "../gas" }
 namada_parameters = { path = "../parameters" }
 namada_storage = { path = "../storage" }
 namada_trans_token = { path = "../trans_token" }
 
 borsh.workspace = true
+lazy_static.workspace = true
 masp_primitives.workspace = true
+masp_proofs.workspace = true
+rand_core.workspace = true
 rayon = { workspace = true, optional = true }
 serde.workspace = true
 smooth-operator.workspace = true
