+++ crates/shielded_token/src/lib.rs
@@ -21,6 +21,7 @@ pub mod conversion;
 mod storage;
 pub mod storage_key;
 pub mod utils;
+pub mod validation;
 
 use std::str::FromStr;
